requirements.txt
lxml==5.2.2
numpy==1.26.4
pandas==2.2.2
SPARQLWrapper==2.0.0
tqdm==4.66.4
yfinance==0.2.38
rdflib~=7.0.0
python-dateutil~=2.9.0.post0
click~=8.1.7
torch
torchvision
torchaudio
pycountry
jupyter